requirements.txt
anyio==4.3.0
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
python-telegram-bot==21.1
rsa==4.9
sniffio==1.3.1
Telethon==1.35.0
